requirements.txt
pandas==2.2.0
streamlit
pillow==10.2.0
pix2tex==0.1.2
pyperclip==1.9.0
plotly==5.20.0
texsoup
st-copy-to-clipboard